:chore : fix: 평가자 페이징 맥스 설정

+++ backend/focus/src/main/java/com/bb/focus/api/service/CompanyAdminServiceImpl.java
@@ -176,7 +176,7 @@ public class CompanyAdminServiceImpl implements CompanyAdminService {
     if (process == null) {
       return null;
     }
-    int cur_step = process.getCurrentStep();
+    Byte cur_step = process.getCurrentStep();
     String processName = process.getName();
     List<ApplicantPassLog> applicantPassLogsList = applicantPassLogRepository.findByProcessNameAndStep(
         processName, cur_step);
